style: 종류에 따른 상위 패키지 생성
