+++ src/main/java/com/c9pay/storeservice/data/dto/product/ProductDetails.java
@@ -1,4 +1,4 @@
-package com.c9pay.storeservice.dto.store;
+package com.c9pay.storeservice.data.dto.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,8 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class StoreDetails {
-    private Long id;
+public class ProductDetails {
+    private long id;
     private String name;
+    private int price;
 }
