+++ src/main/java/com/c9pay/storeservice/data/dto/sale/ProductPurchaseInfo.java
@@ -1,4 +1,4 @@
-package com.c9pay.storeservice.dto.sale;
+package com.c9pay.storeservice.data.dto.sale;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,9 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ProductSaleInfo {
+public class ProductPurchaseInfo {
     private long productId;
-    private String name;
-    private int price;
     private int amount;
 }
