--- src/main/java/com/c9pay/storeservice/dto/QRInfo.java
@@ -1,15 +1,0 @@
-package com.c9pay.storeservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class QRInfo {
-    private String certificate;
-    private String content;
-}
