+++ src/main/java/com/c9pay/storeservice/mvc/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package com.c9pay.storeservice.repository;
+package com.c9pay.storeservice.mvc.repository;
 
-import com.c9pay.storeservice.entity.Product;
+import com.c9pay.storeservice.data.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
