+++ src/main/java/com/c9pay/storeservice/mvc/service/ProductService.java
@@ -1,9 +1,9 @@
-package com.c9pay.storeservice.service;
+package com.c9pay.storeservice.mvc.service;
 
-import com.c9pay.storeservice.dto.product.ProductDetails;
-import com.c9pay.storeservice.entity.Product;
-import com.c9pay.storeservice.entity.Store;
-import com.c9pay.storeservice.repository.ProductRepository;
+import com.c9pay.storeservice.data.dto.product.ProductDetails;
+import com.c9pay.storeservice.data.entity.Product;
+import com.c9pay.storeservice.data.entity.Store;
+import com.c9pay.storeservice.mvc.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
